pyproject.toml
[tool.poetry]
name = "speedy"
version = "0.1.0"
description = ""
authors = ["Lucas Simonet <simonetdaunizeau.l@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
starlette = "^0.37.2"
jinja2 = "^3.1.4"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
huggingface-hub = "^0.23.4"
llama-cpp-python = "^0.2.79"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
